[ShaderWriter API change] Image and texture access rework. (#64)

+++ scripts/ShaderWriter/genFunctionHeaderFromEnum.py
@@ -397,7 +397,7 @@ def getImgTexAccessFormats( depth, sampled, retType, intrinsicName ):
 				formats.append( ( 'R16', retType ) )
 			else:
 				formats.append( ( 'R32', 'Float' ) )
-				formats.append( ( 'R16', 'Half' ) )
+				formats.append( ( 'R16', 'Float' ) )
 		elif intrinsicName.find( "Size" ) != -1 or intrinsicName.find( "Samples" ) != -1 or intrinsicName.find( "Query" ) != -1 or intrinsicName.find( "Gather" ) != -1:
 			formats.append( ( 'Rgba32', retType ) )
 			formats.append( ( 'Rgba16', retType ) )
